+++ gui_widgets/selection_image.gd
@@ -50,12 +50,12 @@ func _connect_selection_manager(_dummy := false) -> void:
 func _gui_input(event: InputEvent) -> void:
 	if event is InputEventMouseButton and event.pressed and event.button_index == BUTTON_LEFT:
 		# image click centers and "levels" the target body
-		IVGlobal.emit_signal("move_camera_to_selection_requested", _selection_manager.selection_item,
+		IVGlobal.emit_signal("move_camera_to_selection_requested", _selection_manager.selection,
 				-1, Vector3.ZERO, Vector3.ZERO, -1)
 
 
 func _update_image() -> void:
-	if !_selection_manager.has_item():
+	if !_selection_manager.has_selection():
 		return
 	hint_tooltip = tr(_selection_manager.get_name()) + _hint_extension
 	var texture_2d := _selection_manager.get_texture_2d()